requirements.txt
Flask==3.0.3
gunicorn==22.0.0
requests==2.32.3
icalendar==5.0.12
python-dateutil==2.9.0.post0